Moving the function from the utils module.
